==> 08_tensorflow.ipynb <==
import numpy as np
import pandas as pd

import plotly.express as px
import plotly.graph_objects as go

from sklearn.metrics import mean_absolute_error, r2_score

import tensorflow as tf

SEED = 1660

df_train = pd.read_parquet("datasets/final_train.parquet")
df_test = pd.read_parquet("datasets/final_test.parquet")

X_train = df_train.drop(columns=["GameID", "Elo"])
Y_train = df_train["Elo"]

X_test = df_test.drop(columns=["GameID", "Elo"])
Y_test = df_test["Elo"]

mean, std = X_train.mean(), X_train.std()
X_train = (X_train - mean) / std
X_test = (X_test - mean) / std

mean, std = Y_train.mean(), Y_train.std()
Y_train = (Y_train - mean) / std
Y_test = (Y_test - mean) / std

X_train.shape, X_test.shape

model = tf.keras.models.Sequential([
    tf.keras.layers.Dense(units=10),
    tf.keras.layers.Dense(units=1)
])

model.compile(
    optimizer=tf.keras.optimizers.Adam(0.05),
    loss='mean_absolute_error'
)

history = model.fit(
    X_train, Y_train,
    epochs=2,
    validation_split=0.2
)

history_df = pd.DataFrame(history.history).assign(epoch=history.epoch).set_index('epoch')

px.line(
    history_df,
    template='plotly_dark'
)

Y_hat = model.predict(X_test)[:, 0]

r2_score(Y_test, Y_hat) * 100
